Refactor according to PR suggestions
* Allow tracer option in span/trace decorator to override default tracer

    * In the tests we're using config of OtherTracer for app :decorator
    to not get a config warning saying we don't have any app named e.g.,
    :other (since :decorator we do have, it's a dependency — check
    mix.exs)

* Renamed Spandex.TraceDecorator to Spandex.Decorators

* Update README

    * Updated instruction on how to override default tracer when
    annotating a method in a given module

    * Better copy following Greg Mefford's texts
