+++ README.md
@@ -198,30 +198,40 @@ Each process should have its own store for its own generated spans. This should
 be fine because you can send multiple batches of spans for the same trace
 separately.
 
-## Decorators (Beta)
+## Decorators
 
-:warning: This is an experimental feature. In the past, we have [deprecated](https://github.com/dnlserrano/spandex/commit/32cf69891d1440cbbcb24b27578cb99a0591cc52) support for decorators. We are now trying to bring them back as part of Spandex's core. Please, report any issues you may find with it.
+Because the  `decorator` library can cause conflicts when it interacts with other dependencies in the same project, we support it as an optional dependency. This allows you to disable it if it causes problems for you, but it also means that you need to explicitly include some version of `decorator` in your application's dependency list:
 
-Configure the Spandex decorator with your tracer:
+```elixir
+# mix.exs
+
+defp deps do
+  [
+    {:decorator, "~> 1.2"}
+  ]
+end
+```
+
+Then, configure the Spandex decorator with your default tracer:
 
 ```elixir
 config :spandex, :decorators, tracer: MyApp.Tracer
 ```
 
-Span function decorators take an optional argument which is the attributes to update the span with.
+Span function decorators take an optional argument which is the attributes to update the span with. One of those attributes can be the `:tracer` in case you want to override the default tracer (e.g., in case you want to use multiple tracers).
 
 IMPORTANT If you define multiple clauses for a function, you'll have to decorate all of the ones you want to span.
 
 ```elixir
 defmodule TracedModule do
-  use Spandex.TraceDecorator
+  use Spandex.Decorators
 
   @decorate trace(service: :my_app, type: :web)
   def trace_me() do
     span_1()
   end
 
-  @decorate span()
+  @decorate span(name: "span_1")
   def span_1() do
     inner_span_1()
   end
@@ -232,9 +242,9 @@ defmodule TracedModule do
     inner_span_2()
   end
 
-  @decorate span()
+  @decorate span(tracer: MyApp.OtherTracer)
   def inner_span_2() do
-    "this produces the span stack you would expect"
+    "this produces a span stack to be reported by another tracer"
   end
 
   # Multiple Clauses
@@ -245,16 +255,16 @@ defmodule TracedModule do
 end
 
 defmodule ThirdPartyApi do
-  use Spandex.TraceDecorator
+  use Spandex.Decorators
 
   @decorate span(service: :third_party, type: :cache)
   def different_service_call() do
-
+    ...
   end
 end
 ```
 
-Note: Decorators don't magically do everything. It often makes a lot of sense to use `Tracer.update_span` from within the function to add details that are only available inside the function.
+Note: Decorators don't magically do everything. It often makes a lot of sense to use `Tracer.update_span` from within your function to add details that are only available inside that same function.
 
 ## Ecto Tracing
