+++ test/support/other_tracer.ex
@@ -1,0 +1,4 @@
+defmodule Spandex.Test.Support.OtherTracer do
+  @moduledoc false
+  use Spandex.Tracer, otp_app: :decorator
+end
